UPSTREAM-REJECTED: src/strategy/actions/foreachpolarisationaction.h: _changeRevised to true
Without setting _changeRevised to true, the "Background" image in the rfigui
never gets updated if the strategy includes a ForEachPolarisationAction. Without
this it's very hard to understand if the iterative algorithm is working.

*** Rejected upstream since this affects the speed of non-GUI operation,
apparently. It was suggested to auto-set _changeRevised based on whether the
GUI is active; apparently other aspects of the strategy autogen already do
things along these lines. ***